+++ tests/unit_tests/test_config.py
@@ -20,12 +20,7 @@
 import unittest
 from unittest import mock
 
-from croud.config import (
-    Configuration,
-    IncompatibleConfigException,
-    load_config,
-    write_config,
-)
+from croud.config import Configuration, load_config, write_config
 
 
 class TestConfiguration(unittest.TestCase):
@@ -40,7 +35,7 @@ class TestConfiguration(unittest.TestCase):
 
     @mock.patch("yaml.dump")
     def test_write_config(self, mock_yaml_dump):
-        config = {"env": "dev", "token": "eyJraWQiOiIxZTlnZGs3IiwiYWxnIjoiUlMyNTYifQ"}
+        config = Configuration.DEFAULT_CONFIG
         m = mock.mock_open()
         with mock.patch("croud.config.open", m, create=True):
             write_config(config)
@@ -49,9 +44,17 @@ class TestConfiguration(unittest.TestCase):
             config, m(), default_flow_style=False, allow_unicode=True
         )
 
-    def test_incompatible_schema(self):
+    @mock.patch("croud.config.os.remove")
+    @mock.patch("croud.config.os.chmod")
+    @mock.patch("croud.config.write_config")
+    def test_incompatible_schema_rewrites_config(
+        self, mock_write_config, mock_chmod, mock_remove
+    ):
         config = {"incompatible": "", "schema": ""}
-        with self.assertRaises(IncompatibleConfigException) as cm:
-            Configuration.validate(config)
 
-        self.assertTrue("Incompatible storage format" in str(cm.exception))
+        m = mock.mock_open()
+        with mock.patch("croud.config.open", m, create=True):
+            Configuration.validate(config)
+            self.assertTrue(mock_remove.called)
+
+            mock_write_config.assert_called_once_with(Configuration.DEFAULT_CONFIG)